crawler/requirements.txt
selenium==4.38.0
webdriver-manager==4.0.2
google_genai==1.4.0
python-dotenv==1.2.1